refactor: YAGNI audit — remove dead code, wire lying config fields, fix stale docs
Code cleanup (9 findings):
- Remove duplicate __init__ assignments in Pretrainer (pretrain.py:217-219)
- Wire ModelConfig.init_std into _init_weights instead of hardcoded 0.02
- Wire YAML beta1/beta2 into TrainingConfig (were defined but never read)
- Delete dead ModelConfig.dtype field + all 11 references (YAML, README, tests)
- Delete dead imports in transformer.py (math, Optional, torch.nn.functional)
- Drop dead d_model param from MIMO.__init__ + update all call sites
- Simplify redundant except tuple in mamba_block.py to bare Exception
- Rewrite stale SKILLS.md (referenced non-existent files, superseded rules)
- Fix AGENTS.md §4 file list (remove inference/, documentation/; add actual files)

Tests: 28 passed, 1 skipped, 0 failures.

+++ tests/e2e_gpu_smoke.py
@@ -34,7 +34,7 @@ def _tiny_cfg(ssd_dispatch: str = "pytorch") -> ModelConfig:
     return ModelConfig(
         vocab_size=128, d_model=64, n_layers=2, n_heads=4,
         head_dim=16, state_dim=16, chunk_size=16, ffn_dim=128,
-        max_seq_len=32, dtype="fp32", weight_tying=True,
+        max_seq_len=32, weight_tying=True,
         ssd_dispatch=ssd_dispatch,
     )
 
@@ -275,7 +275,7 @@ def check_pretrainer_dry_run(device: torch.device):
             model_config=dict(
                 vocab_size=128, d_model=64, n_layers=2, n_heads=4,
                 head_dim=16, state_dim=16, chunk_size=16, ffn_dim=128,
-                max_seq_len=32, dtype="fp32", weight_tying=True,
+                max_seq_len=32, weight_tying=True,
                 ssd_dispatch="triton", grad_checkpoint=False,
             ),
             data_path=str(shard_path),
